Amazon Product Analysis
Hi,

Thank you for checking my project for today which is "Amazon Product Analysis". 
Mainly, my analysis is based on the "Actual Price" of the product.

I have struggled with the database, as it was a huge dataset of more than 500K rows which was very exhausting to clean the data, preprocess the data..ETC

I had some challenges with getting the brand name, but I extracted the name of the brand by using the name column. 



After digging deep into the data, we arrived at the following conclusions:

1- The products with prices less than ₹3600 are popular.
2- The outlier in price data is around 5.32%
3- Puma is the most popular brand with an outlier price  
4- Character and brand is the subcategory with outlier price
5- The maximum number of ratings of popular brands is in the range of 4-star
6- Mostly 0-49 reviews were given on the products
7- Every product has a review in the selected dataframe
8- Distribution of products by discount percent shows distribution similar to normal with slight right skew and spikes at each value multiple of ten We can analyze other relationships as well but for now the above is the conclusion.

The Project took me 2 days to complete, and there is a lot of insights that data can provide to us.